index.py

+++ client.py
@@ -12,7 +12,7 @@ async def send_audio(uri):
             # Send the audio data asynchronously
             asyncio.run_coroutine_threadsafe(websocket.send(indata.tobytes()), loop)
 
-        with sd.InputStream(samplerate=16000, channels=1, callback=callback):
+        with sd.InputStream(samplerate=16000, channels=1, dtype='int16', callback=callback):
             print("Start speaking...")
             while True:
                 try:
@@ -24,6 +24,7 @@ async def send_audio(uri):
                 except Exception as e:
                     print(f"An error occurred: {e}")
                     break
+
 
 # Run the asyncio event loop
 asyncio.run(send_audio("ws://127.0.0.1:8000/ws/transcribe/"))
